delete the old prepare corpus and rename preprocess_avi to prepare_corpus
